pyproject.toml
[project]
name = "agent-service-toolkit"
version = "0.1.0"
description = "Toolkit for running an AI agent service built with LangGraph and FastAPI"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"

dependencies = [
    "fastapi ~=0.115.5",
    "grpcio >=1.68.0",
    "httpx ~=0.27.2",
    "jiter ~=0.8.2",
    "langchain-core ~=0.3.33",
    "langgraph ~=0.3.5",
    "langgraph-checkpoint-sqlite ~=2.0.1",
    "langgraph-checkpoint-postgres ~=2.0.13",
    "psycopg[binary,pool] ~=3.2.4",
    "pydantic ~=2.10.1",
    "pydantic-settings ~=2.6.1",
    "python-dotenv ~=1.0.1",
    "setuptools ~=75.6.0",
    "uvicorn ~=0.32.1",
    "langfuse>=2.60.3",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

# Group for the minimal dependencies to run just the client
# To install run: `uv sync --frozen --only-group client`
client = [
    "httpx~=0.27.2",
    "pydantic ~=2.10.1",
    "python-dotenv ~=1.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
# No test environment variables needed for LLMs

[tool.mypy]
plugins = "pydantic.mypy"
exclude = "src/streamlit_app.py"

[[tool.mypy.overrides]]
module = ["numexpr.*"]
follow_untyped_imports = true
